Add timer and make things pointers

+++ gobotcore/util.go
@@ -2,9 +2,9 @@ package gobotcore
 
 import "strconv"
 
-func AtoiEZPZ(str string) int {
+func AtoiEZPZ(str string) int8 {
 	if i, err := strconv.Atoi(str); err == nil {
-		return i
+		return int8(i)
 	}
 	return -1
 }
